requirements/base.txt
django
django-environ
pillow
django-ckeditor
django-imagekit
django-modeltranslation
python-slugify
clickuz
djangorestframework
payme-pkg
django-rosetta
gunicorn
psycopg2-binary
django-cors-headers
pytest-django
python-telegram-bot
PaycomUz